Simplify single_actor_batch crash fix on sim_t::cancel
Introduced with f24fc9d, this commit instead ensures that no work is
going to be done, if the simulation is cancelled. This ensures that
work_queue_t::pop will not increase the index past the number of actors
in the sim, which in turn leads to crashes.

Generally, the iteration code should never need to explicitly check
bounds for anything related to "current_index" (current actor being
simulated). By definition, if current_index is past the number of
actors, then the simulation work is finished.