Implement some basic consumable options
- New options 'potion', 'flask', and 'food' added as player-scope
options. Each defines the "default consumable" for the actor for the
consumable type
- A special consumable option value 'disabled' will denote that the
consumable type is disabled (the corresponding action does nothing). Old
allow_X disable options still work.
- Add player_t::default_(potion|flask|food) overridable methods to return
a default consumable of such type for the actor. See sc_shaman.cpp for
examples.
- Potion, food, and flask actions now do not require a 'name' or 'type'
parameter. If omitted, the default consumable of the type is used
- A new consumable action expression (i.e., functional on potion, food,
or flask APL lines) 'consumable.X' added. Evaluates to 1 if the
corresponding consumable type is set to X.