Adjust how "drop level" items parsed from armory are initialized
Blizzard armory API does not provide enough information to fully
determine the item level of most quest and world-drop items looted
before the player hits 110. These items require the "drop level"
information, that is correctly provided by other sources (i.e., simc
addon can parse it from itemlinks).

This commit changes the armory import so that the drop level of such an
item is always set to the player's current level. The user will be
notified of this kludge per item, and it will in some cases result in
slight overvaluation of such items.