Some CLI progressbar changes
- Now shows phases properly, i.e., stat scaling identifies on the
progressbar what stat is being scaled, and what actor is it being scaled
for
- Similar to above, plot stats and reforge plot stats are being
identified per actor
- Add a new option "progressbar_type" that adjusts the progressbar
output. Current default is 0, when set to 1, will change to a more
parseable form of a progressbar. The progressbar outputs a single
tab-separated list of fields per line. Fields and their presence are
given below:
  1. Type of simulation (baseline, stat name for stat scaling, stat
  delta for plots, etc.)
  2. Actor name (only on single_actor_batch=1)
  3. Current iteration
  4. Total iterations
  5. Current mean (only when target_error is used)
  6. Current error (only when target_error is used)
  7. Estimated remaining time for simulation (per character when
  single_actor_batch=1)
  8. Simulation time (only when simulation finishes, per character when
  single_actor_batch=1)