Introduce "use_items" action
The intent of use_items is to ease on-use item usage for uesrs. By
default, it will scan through the items of the actor during
initialization and potentially create a "proxy use_item" action for each
usable item. The goal is that use_items can react to item changes (e.g.,
change a trinket), without having to touch the APL for the actor to use
that item.

Use_items tries to be very cautious. It will only create a proxy action
if:
- There is no "use_item,slot=X" action in any APL for the item slot
- There is no "use_item,name=X" action in any APL for the item

Use_items will use all available item effects in order, starting with
the trinket slots (trinket1, trinket2), followed by the rest of the
slots. Slots can also be customized by passing use_items the "slots"
option, in which case that specific use_items APL line will only try to
use the slots specified in the option value. Multiple slots are
delimited by : or | characters. Normal "if" conditions and such apply to
use_items actions.

Experimental, may contain bugs, etc.

This commit also fixes a very rare bug where on-use based buffs in some
rare cases might not trigger when the item is used.