Add a workaround for weird threading behavior on new windowsen
At least Windows 10 scheduler treats threads in a peculiar way, where
the main thread of the simulator is given disproportionately little time
to do it's job, compared to other threads in the process.

This leads to main thread not being able to do some critical jobs it is
required to perform, such as target error computation, resulting in very
very odd behavior while simulating with the command line client.