Add support for expiration pulses to ground_aoe_event_t
ground_aoe_params_t::expiration_pulse sets the mode, defaults to no
expiration pulse. Other options include FULL_EXPIRATION_PULSE for a
full-power action execute on expire, and PARTIAL_EXPIRATION_PULSE for a
relative-power action execute on expire, based on the time left between
the penultimate tick and the duration of the aoe event, and the computed
pulse time.