Initial pass on Cradle of Anguish
Also adds a sim-scope option "legion.cradle_of_angiuish_resets" that
takes a list of time values (delimited by comma, :, or /) when the
Cradle of Anguish buff is expired (meaning the actors went below 50%
health). Defaults to never, meaning Cradle of Anguish is modeled as a
stat stick.

Note that the "must be over 80% health to gain stacks" is currently not
modeled.