[Hunter] use 'multishot' as Multi-Shot token name
Instead of 'multi_shot'.

We currently have a bit of a problem because of a discrepancy between
the token name given to the action ('multi_shot'), the name put into apl
('multishot') by the tokenizer and the names used in apl conditions
(both of them).

This does not matter for 'standard' sims but with distance targeting
enabled this breaks somewhat.

Standardize on 'multishot' mostly because that's what the tokenizer
gives us.