[Hunter] drop Hunter's Mark rppm manual "hotfix"
Essentially revert "[Hunter] Update Hunter's Mark rppm based on tests"
Commit c415fb11f66d8ae684a78d5e243e44a1910401c1.

The pastebin at the link in the commit is long gone, so we can't verify
the numbers from it, but it looks like the testing done there failed to
account for the increased real count of procs due to bad luck protection
inherent to rppm.

Some preliminary testing shows that 6.5ppm in the spell data is actually
correct.

Real life testing:
https://www.warcraftlogs.com/reports/xnq7thP9XvaW3B1Y/#view=analytical&type=auras&source=6
2 runs at 1287 (951 + 336) procs in ~175 (130 + 45) minutes, ~7.35ppm

Running the same in simc:
https://www.raidbots.com/reports/B1QO4wn7-/index.html
1406 procs, ~8.03ppm

Using the spell data 6.5ppm on Hunter's Mark:
https://www.raidbots.com/reports/HkwKVv2XZ/index.html
1273 procs, ~7.28ppm

While ideally we'd get a longer contionuous test it's already pretty
obvious that our current "hotfix" value of 7.2 is wrong and 6.5 from the
spell data looks like the actual ppm.

Does not really affect the absolute simmed dps much (the diff is almost
completely in the error range) but increases the dps range by ~10% (from
~270k to >300k for T20M profile).

Big thanks to @vituscze for bringing this to my attention.