[Hunter] rework summon_pet option
Implement a real default that's active even with a custom APL.

We currently have a bit of a problem with "custom" apls where you *have
to* also have a 'summon_pet' option present because otherwise it's empty
and the hunter can't summon a pet. This is somewhat non-obvious.

On another note we can actually completely rip it out now as pet family
does not matter anymore.

And on another note we maybe should give the main pet a set name
('fluffy'?) so that the apl is pet family agnostic.