[Hunter] add Bestial Ferocity dot expression to Kill Command
Unfortunately there is no real way to make working expressions involving
pets dots atm (even pet.dot.xxx does not really work as it uses the
current apl *action* to get the player, which is the owner). But we
kinda need it for Aspect of the Beast aoe rotation to spread the dot
around with cycle_targets.

Add a special expression 'bestial_ferocity.ticking' to Kill Command to
make it possible.