Fix an issue with channeled spells and chain/interrupt based APLs
Channeled spells need to perform resource regeneration on tick, if
there's an interrupt or a chain based option passed on the APL line.
This is because both of those systems require up to date (to the
millisecond) information for the current resources to make the correct
decision.

This will fix for example Affliction Warlock behavior, where the action
sequence table had two consecutive Drain Soul actions for no apparent
reason.