[Priest] Decouple death, taxes, and s2m
Surrender to Madness kills the prices, which can occur in unfortunate
places in the simulator execution path. This commit moves the killing of
the actor to a discrete event that occurs on the same timestamp as the
"source event" that caused the death (i.e., out of insanity).

Fixes a rare(ish) crash where a mind-flay tick may trigger s2m death.