Initial sketch of the Antorus empowered pantheon bonus system
The empowered pantheon trigger system is modeled as a combination of
proxy users based on user input and the items carried by actors defined
in the simulation profile. Each "proxy user" represents another actor in
the environment who owns the type of trinket specified and attempts to
proc it. Internally, each proxy user is represented by a discrete
real-ppm object to simulate the proxy user attempting to proc their
representative trinket throughout combat.

The system tracks both the proxy and actual user trinket states, and
upon seeing all pantheon auras up, triggers the empowered trinket
effect(s) on all actors owning a pantheon trinket. Currently, the proxy
user system attempts to trigger the proxied trinkets every second.  This
may change in the future.

This initial commit introduces the pantheon state system, including the
proxy users, and the empowered effect of Aggramar's Conviction.

A new sim-scope option "legion.pantheon_trinket_users" has been added to
specify the proxy users for your simulation environment. The format of
the option is a '/'-delimited set of tokens of the form:
<type>:<haste%>, where <type> is one of 'am', 'go', 'kh', 'eo', 'no',
'ag', for Aman'Thul's Vision, Golganneth's Vitality, Khaz'goroth's
Courage, Eonar's Compassion, Norgannon's Prowess, and Aggramar's
Conviction, respectively. The haste% value is the paperdoll haste value
of the proxy caster.

For example, the option
"legion.pantheon_trinket_users=am:0.1/am/go/eo/no/ag" would have proxy
users for all but Khaz'goroth's Courage, with two Aman'Thul's Vision
users, out of which one has a static haste level of 10%. The intent here
would be that the user's own actor would be handling the Khaz'goroth's
Courage proccing.

For now, there are no default proxy casters in the environment, so there
will be no empowered buffs triggered. This is likely to change in the
future.