Merge Time: Change to measure only main thread merge time.
Previously this measured the time of all threads (mostly of threads
waiting on lock). I think we are more interested in the total sequential
merge time.
Or if that is not what is intended @navv1234 , I think we should at
least move the start of the time-measurement after obtaining the lock.