Implement multi-metric gathering to profilesets
The profileset_metric parameter now accepts a '/', ':', or ',' delimited
string for multiple different metrics to gather from each profile. The
output is currently only given in the 'json2' option, where a new object
"additional_metrics" can be found for the individual profileset result
objects.

The first defined profileset metric is considered the "default"
profileset metric, and displayed in the HTML and text reports (and as
the main profilset metric entry in the JSON report. If no profileset
metrics are defined, the default metric is DPS.