[Shaman] Update Storm Tempests modeling to recent findings
At some point Blizzard seems to have changed the legendary to
immediately zap all targets upon debuff application, and also made it so
that the debuff is applied twice (each damaging strike). In addition to
this, there is some yet-to-be-determined behavior occurring with the
refreshing that causes the actual debuff duration to be longer than the
client-data derived 15 seconds. Simulationcraft behaves so that the
initial application (target has no debuff) adds 2 extra seconds (ticks)
to the duration, and each refresh resets the debuff duration to 16
seconds (a single extra tick).