Implement (more) proper school-based value computation for spells
Previously, the sim would pass the school of a spell to the various
multiplier and spell power functions to determine the values used for
amount computation.

In game behavior actually uses the maximum base school-based value for
the attribute. As it stands there are no spells in normal simulation
where this would be relevant, but it is better to have the core behave
correctly in the case ever arises.

This change has been applied to player and target-based multipliers,
spell power, and persistent multipliers.

In simc, this is technically implemented by breaking down spells to
their base schools, and calling the relevant methods with each of the
base schools. In addition if the spell is using a composite school
(e.g., Astral), a separate invocation to the relevant methods is done
with the composite school. This is to support a few special cases, where
it is relevant to determine the relevant multipliers using the composite
school.