Initial implementation of parallel profileset processing
Previously, profilesets were processed sequentially, using the same
number of threads as the parent (baseline sim) was simulated with.

This commit introduces a new profileset processing mode, where multiple
profilesets can be processed in parallel. The new
'profileset_work_threads' defines the number of threads (out of the
simulator maximum threads) can be used to process a single profileset
simulation. For example, profileset_work_threads=1 would use a single
thread for each profileset, allowing up to threads= number of parallel
profilesets to be processed. The maximum number of threads is
constrained by the normal 'threads' option.

Some rough edges here, as this is the initial commit.

1) The progress bar is very confused, this will be fixed at some point
2) Interrupting (ctrl-c/sigint) sims may get stuck in some rare cases
3) Error handling is very basic at the moment

Buyer beware.