Remove Initial Chi comments
It's been decided that we stick with zero chi at the start of a simulation instead of setting it to 1 chi. Chi gets set to 1 chi if the player goes into the fight with 1 or more chi. Blizzard drops us down to 1 chi to avoid abuse. However it can be difficult to be at 1 chi going into the fight consistently.