[Hunter] initial BfA Marksmanship update
Apart from possible bugs the main not implemented thing is Trick Shots
as it doesn't work on dummies so pretty hard to test and it's also not
very obvious how to properly implement it simc-wise.

All talents should be working, even Volley (with the proc chance taken
from the talent).

APL is obviously very preliminary.