8.0.1.26476
Note that this build brings some changes that affect simc

1) The old style effect "random" ranges have been removed. The concept
of "die sides" is gone, replaced with the newer style base value + delta
coefficient. This is mostly an invisible change, however Demon Hunters
in particular used the die sides system for randomized resource
generation. This needs to be changed.

2) Effect base values change changed from a 32 bit integer value, to a
floating point value. This should not affect anything.