Disable automatic base damage calculation for power-scaled spells
Previously, actions unconditionally set the base damage
(base_dd_min/base_dd_max/base_td) to a (potentially) scaled value based
on spell data. Now, actions will only set the base value if there is no
power-based scaling defined for spells.

Reasoning for this is that base damage for the vast majority of spells
in WoW has not existed for quite a few expansions. It is time to default
to "no base damage" for abilities, unless very specific conditions
exist, instead of defaulting to "always compute a base damage based on
client data, regardless of if it exists or not".