[Hunter] transform Dire Frenzy into Barbed Shot
This still doesn't really work how it says on the tooltip and half of
the stuff is either broken or not updated, but it looks like it really
is the direction they are going to push it.

It also contains some updates to Dire Beast as they are not exclusive
anymore.

Let's hope they don't change it back.