Expressions: Decouple player/sim/buff expressions from actions.
Expressions for players, the sim and static buffs are conceptually no
longer "action-expressions", but only expressions about the object they
refer to.

This allows for the second part of this commit:
Add player_if= expression to raid_events to allow the expression system
to be used as a filter for the affected_player list of a raid event.