[Hunter] fix up Rapid Fire tick count
The spell actually has the total number of ticks in the spell data with
all of the spells affecting its duration modifying it at the same time.
This is important because it has a zero tick, which means that atm it
almost always has 1 more tick than it should. It also means that in-game
it actually never channels for the full 3 seconds even with 0 haste.

This is a bit more complicated than it could've been because Double Tap
is snapshot on the start of the channel. Which requries adding a custom
action_state to make it consistent.