[Hunter] adjust Rapid Fire a bit
Adjust the stats & setup of the child action to mimic what tick_action
does to simplify the reporting.

Unfortunately we can't use tick_action for the periodic action as the
core uses schedule_execute() for the former which is incompatible with
Trick Shots as the last tick gets executed after the main actions
last_tick().