Don't allow automatic inference of legendary special effects
In some cases, the automatic inference of special effects that generates
simple trinket procs for us automatically, will incorrectly create a
special effect proc for legendary items. This is because the client
spell data is incomplete, and for the inferrence system, it looks
"usable", where in reality it is not.

An example would be the Warlock legendary "Recurrent Ritual". The client
data lacks information about which spell can trigger the effect, so the
automatic inferrence thinks that it is triggerable by all spells.