cmake improvements (integrate into synchronize.py) (#4347)
* cmake: Fix qt build for macOS and newer qt5 versions

In addition update qt configuration to new format:
	- WebKit no longer available (only WebEngine)
	- Use new CONFIG instead of COMPONENT style discovery
	- People with old qt5 can still use the Makefile instead of cmake

* cmake: Add SSL library detection

* cmake: Move git-hash code to engine CMakeLists.txt

It effects an engine *.cpp file that provides the hash
to its users and should therefore reside in the corresponding
CMakeLists.txt. This helps us in maintining a clean root
CMakeLists.txt.

* synchronize.py: Change missleading comments in *.pri files

Also prevent now obsolete code in synchronize.py that oberverwites
.pri files with its own contents. This is not neccessary as these
are the files we edit by hand.

* synchronize.py: Generate source file lists for cmake

Before this patch the whole cmake file was overwritten on
synchronize.py. Now we generate a source file list for inclusion
just like for the other build systems and avoid globbing for cmake.

* cmake: Set C++ standard to C++11