Add mother event and misc
adding mother on mythic and hc difficulty. Since mother is a very odd fight I had to split it into two seperate events. The user will use one of the two depending on his job in the raid team.

Misc is fixing a few names, naming some other events and adding in bloodlust dummies for Fetid Devourer and removing races that are yet to be implomented on bosses.