Add Zul
adding in the profiles for Zul on Heroic and Mythic. Need to check the actual numbers again after logs are public from guilds reaching phase 2. Numbers now are based on estimates from phase 1 logs and boss abilities tooltips.

Added a readme to keep track of what still has to be done.