Adjust azerite trait value calculations
There is currently (in game) a mystery way of stating that an azerite
trait that grants ratings in a passive way (For example Blood Siphon) is
automatically penalized by the combat rating penalty (based on ilevel),
even though the typical scaling type for such a spell is -7 (where as
these passive spells use -1).

Since we do not know where in the client data this is indicated,
implement a relatively strict detection method in azerite_power_t to
apply the combat rating penalty. This also likely fixes some off-by-one
errors in some traits, based on some in game testing of displayed values
vs simc values.