Add automatic determination of "spell type" based on client data
This commit changes the way the first-phase determination of whether a
proc should be attempted is changed. Previously, we used simc-internal
logic to determine whether an action is categorized as a spell or melee
ability or white melee, etc. Now, the categorization is automatically
done if an action is associated with valid spell data.