Synaptic Spark Capacitor
Some stuff unknown about this, namely, what is the aoe damage spell that
does the actual damage to "nearby enemies". Need a combatlog with the
power to find the spell id.

Also, the ticking is done in an awkward way (1950ms period), so there's
a final tick since the debuff lasts 10 seconds. Is the final tick a
damaging one? Does it do full damage? Or is the actual length of the
debuff 9750ms instead?