[Azerite] Battlefield Focus / Precision
Note, this implementation is not how it works in game. The Battlefield
debuff is personalized to the actor that procced it, instead of being
global. In addition, the number of stacks on the application of the
debuff is by default restricted to 2 (instead of 20). The number of
debuff stacks can be controlled with a sim-scope option
`bfa.battlefield_debuff_stacks` that takes values in the range 1-25.