[trinket] first pass at Leyshock's Grand Compilation
Does not currently work. Modeled after Harlan's implementation.

I spent some time testing this on Beta, and it appears that all
casts and DOT ticks are capable of proc'ing this, and the proc
rate is fairly high. It behaves similarly to Acrid Catalyst
Injector, if ACI lacked the "all" buff.

This implementation doesn't work quite yet, but I figured I'd throw
this up for review, as someone more familiar with the process of
implementing trinkets will surely have better ideas on how to do
this properly.