[Mage] Allow users to specify the duration of freeze effects.
By default, we assume the freeze effect gets broken instantly and remains
only for the 0.5 sec grace period. In BfA, the roots are more durable and
require a lot more damage to break, so only using the grace period might
underestimate the gain from using these effects.

This adds a character-wide option named 'frozen_duration', which specifies
the duration of freeze effects (in seconds). Setting it to zero or less
makes the freeze effect permanent. As usual, the debuff is only applied to
adds from the add raid event or bosses whose level is less than player
level + 3.