Refactor ability readiness and fix readiness check on cast finish
This commit refactors how the simulator determines the readiness status
of actions. Before, a single action_t::ready() method was responsible
for deciding the status, now three different "ready" methods are
responsible for it.

1) action_t::ready() is responsible for checking for ability inherent
conditions, such as resource cost, certain buffs being up, a talent
replacing the spell and so on.

2) action_t::target_ready() is a new method, and its sole purpose is to
check the readiness of the ability to execute on a given candidate
target. Target_ready() is invoked by action_t::action_reeady() and on
each candidate target when using target_if option.

3) action_t::action_ready() is the new entry point into the complete
readiness of an action (replaces the old action_t::ready()).
Action_ready is responsible for analyzing readiness by calling both
action_t::ready, and action_t::target_ready on a candidate target, and
analyzing the readiness of the APL line, based on user input (most
importantly, based on any if condition).

In addition target selection for each action has been split into a
separate method (action_t::select_target)) that is overridable.

This commit also fixes a few subtle bugs in a couple of abilities, where
the action_t::ready() and/or target_if and cycle_targets were resulting
in a slightly incorrect state. Also fixes a bug in cooldown-based queue
spell casting where the readiness state of the ability was not properly
checked when the spell actually executed (cooldown is elapsed).
Similarly, the simulator now also checks the castability of the
non-instant spell when the casting time has elapsed.

Overall, a slight performance decrease is expected (1-5%), however some
specs may see a slight increase in perforance also. Metric values should
be very close to the old values, as incorrect events fixed by this
commit are relatively rare.