[Arcane] Change APL to not use random arcane blasts during burn phase in AoE situations
The line being changed only exists for the situation where on single-target encounters, the mage ends up with a Ro3 buff (due to using charged up or arcane orb) in the burn phase. The line however also triggered for AoE situations because it is constantly getting rule of threes during the burn phase due to arcane explosion usage.