[DBC] Transition to full wow build versioning
During PTR cycles when live and ptr go out of sync in terms of client
data file, the build number is not enough to discern different versions
for the exporter. Thus, convert the exporter to require full version
numbering for the -b option (i.e., 8.1.0.27934).

The exporter will still support the old build-number -b option until
8.1.0 goes live.

[ci skip]