Specialized version of executing and execute_remains expressions.
This introduces casting and cast_remains that check whether the actor is
currently casting the given action (or how long until the cast is over).
channeling and channel_remains do the same, except for channels. executing
and execute_remains now work with casts and channels, whichever is
applicable.

Also fixes a bug with executing and execute_remains that would prevent it
from working correctly when the APL had multiple lines with the same
action.