Implement support for Blizzard's new api endpoints
This commit also adds libcurl as a requirement for simulationcraft
compilation (unless SC_NO_NETWORKING is used). Instructions on how to
use libcurl on Windows, Linuxes, and MacOS are in the HowToBuild wiki
entry.

With the new api endpoints, new client credentials are also needed. Wiki
BattleNetImport instructions have also been updated to reflect the new
process.

The old "apikey" option now accepts the new client credentials (in a
colon separated format). A new option "apitoken" has been introduced
that allows the user to supply a custom oauth 2.0 client credentials
bearer token to use for the armory api requests.

Simc tries relatively hard to cache bearer tokens and to reuse them for
as long as the api endpoints allow. On windows, the bearer token is
cached in a text file, either in %HOMEPATH%%HOMEDRIVE%\simc-apitoken, or
in the directory simc.exe (or simulationcraft.exe) is invoked from. On
unixes, the default save path is $HOME/.simc-apitoken, or if $HOME is
not available, in the current working directory.

All reasonably sensible build systems that Simulationcraft uses are
updated to support the libcurl compilation.