Variable Intensity Gigavolt Oscillating Reactor, part 1
Very initial implementation of the trinket. Supports the on-equip effect
of the trinket (the oscillation buff). Adds a new sim-scope bfa option
"bfa.randomize_oscillation" (defaults to 1) that specifies whether
the oscillation state starts randomized at the beginning of combat. When
explicitly set to 0, the trinket always starts in the ascending phase at
stack 0 (beginning of combat gives the first stack).

On-use effect support will arrive later.