Variable Intensity Gigavolt Oscillating Reactor, part 2
Introduces the on-use action. Due to how it works, an additional option
has been added to the simulator. A sim-scope option
"bfa.auto_oscillating_overload" controls whether the actor is presumed
to be automatically casting the on-use effect when the oscillation
reaches its peak (i.e., 6 stacks of the buff). The option defaults to 1
(true). Disabling the option means that use_items and use_item actions
take over the control of the on-use effect.

Note that this means that for manual control of the Oscillation
Overload, the option must always be set to false.