Eliminate ItemRandomSuffix from the simulator
Blizzard has stopped using ItemRandomSuffix client data for any kind of
random suffixes in BfA, and the client data file has been empty for a
while. In 8.1.5 the file will be removed from the client, breaking
Simulationcraft.

This commit removes all functionality for the old-style item "suffix"
option, and deprecates the option.