[Death Knight] 8.1.5 Changes Remove PTR checks
Update Magus of the Dead:
The pet now chain casts and has a more reliable AI.
Frostbolt will be used on cooldown (3s not in spelldata, manually registered as a hotfix) and applies a 4s snare on non-boss targets hit. Frostbolt won't be used if the target is affected by that snare (not shared between the two pets)
Contrary to patch notes, the pet can still auto attack, although not as much as before.