Initial implementation of Potion of Focused Resolve
Simc models the "marking" so that the system only remember the last
target hit. This means that on aoe, the mark will rapidly jump through
targets until landing on the last target hit for that specific aoe.

In game testing needed to see the correct behavior with aoe spells.
Also, uses a new effect sub type (306), shared with "Blood of the Enemy"
Azerite Essence. Simc presumes they stack additively.