Very initial implementation of Yellow Punchcard logic
Currently ilevel is hardcoded to 415 for a couple of reasons:
1) Simc will need new magic to support item-level dependant gem scaling.
Thus far everything has scaled based on the actual item itemlevel.
2) At least the Blizzard API does not support conveying any other
information about gems than their item id. This is not sufficient to
fully support the punchcard trinket.
3) Simc addon will need to be updated so that we can extract either gem
bonus ids, or gem ilevel from in-game punchcards, and convey that as an
option associated with that item option