Aquipotent Nautilus initial implementation Assumptions made from tooltip.
Sends out a wave (unknown id) that applies a 4s dot (0tick, crits, nonhasted, id=302580) to all targets hit. If you catch the returning wave cooldown is reduced by 30s. Unknown what the travel speed of the outbound & returning waves are. HARDCODED to be 10yd/s both ways for now.

'bfa.aquipotent_nautilus_catch_chance=X' can be used to set the chance to catch the returning wave. Defaults to 1.0.

NOTE: The driver triggers spell (id=302550) that has a 12s duration. Currently unknown what the purpose of this spell and it's duration is.