Extend "trinket" expression syntax
Experimentally allow expressions of the form "trinket.<trinketname>.x",
where trinketname is the tokenized form of the item name. This will
check the actor's current trinket slots for the existance of the item,
and if found, use that to generate the expression (based on the suffix
string x).

If the item is not found or the item is not a trinket, the expression
will always return 0.