Hack to simulate prepull channeling of Font of Power 'bfa.font_of_power_precombat_channel=X' will set how many seconds before the pull you start channeling. When combat starts, it will apply the appropriate duration buff & cooldowns.
Defaults to 4s, which mean with this build, ALL characters equipped with the trinket will prepull channel for 4s and start combat with a 30s buff.

NOTE: This is just a hack and not a true precombat action, so you may want to adjust precombat apls as necessary when equipped with the trinket.