[Gear] Initial implementation of Font of Power /use_item in precombat APL * 'actions.precombat+=/use_item,name=azsharas_font_of_power' can be added a non-harmful, non-channeled, buff-triggering action to the precombat APL.
This will trigger the buff based on spec-specific prepull channel times, which can be hardcoded per-spec in the precombat_buff() function of items::azsharas_font_of_power().

The default value is 4s, which at the start of combat will give a 30s buff, 116s trinket cd, and 16s shared cd.

* However, if the 'bfa.font_of_power_precombat_channel' option is set, this precombat action will do nothing, and the buff & corresponding cooldowns will be applied following the option settings.