Prefer spell-data based cooldown category for special effects
Blizzard internally started using spell-driven cooldown categories a
long while ago for item effects, instead of the "item effect"-based
information. As a result, there is somewhat odd data in the game, where
the item-effect supposedly has a category cooldown, however the category
cooldown duration is set to a bogus (or special) value.