Fix ticking-related issues in buffs. Implement buff partial tick support
Buff ticking could end up in a state where the buff would not actually
tick if it was refreshed inside the period between the penultimate tick
and the expiration (tick). Similar case also affected positive duration
extension of ticking buffs.

Fixing the above issues also results in buff partial ticks (i.e., an
additional tick at the end of the buff duration that has a tick time
lower than the amplitude of the periodic effect). Since old behavior was
to not allow partial ticks, there is now a new state that can be set on
buffs with buff_t::set_partial_tick( true/false ). The default value of
the flag is set to false.