Adjust profilset option parsing logic
Due to the simplicity of the simc option parsing system, profilesets
figure out a position for the "profileset option set" in the profileset
sim by looking for certain options in the base option set. Due to this,
it is possible to organize the combination of profileset and base
options such that the profileset specific options do not actually
override identical options in the base option set.

This commit adjusts the behavior such that if the system considers an
option overridable, the profileset option handling will actually
override the base option value, instead of inserting it to a specific
position in the combined set of options.