[Shaman] Add Elemental pet related expressions to the module
With syntax "pet.(primal_|)(fire|earth|storm)_elemental.(remains|active)

The "primal" prefix will force the elemental type to be explicitly the
primal version, i.e., if the actor does not have Primal Elementalist
talented, the expressions will return their "false" equivalent value.

The `active` expression will return 0.0 if the selected elemental is not
active, and 1.0 otherwise. The `remains` expression will return the
remaining duration of the selected elemental in seconds, or 0.0 if the
elemental is not active.