[Shaman] "Fix" Glimpse of Clarity for Stormbringer procs
A more generic fix requires core changes, and some thinking on the
nature of execution state in general. As it stands, the change where
proc execution became a discrete event (instead of happening during
execute event) causes some issues for the ability execute model (in
class modules typically), since state changes during execute will affect
the proc. In the previous proc execution system, this was not the case,
unless explicitly coded for.