[DBC] pseudo-runtime spell driver linking
Generate per-spell spell drivers "arrays" during spell data export
generation.

One caveat is that we have to revert exported spell data `std::array` to
a c-style array because `std::array::operator[]` for non-const arrays is
constexpr only in C++17 and we really need constexpr here to force
compile-time generation of pointer array.

Otherwise this is the same as was done for powers & labels.