Sort source file lists with python
Instead of relying on the order "given" by the globbing routine which is
filesystem specific and thus can't really be relied upon. Case in point are
the 2 dbc headers at the end of the list.

While it does not really matter (the lists are ultimately the same) it lowers
the potential diff churn.