Switch most of util to string_views
Some functions here are called with string literals a lot, so it's nice to skip
constructing the temporary strings. It's also the easiest to switch as it's all
"leaf" functions.

NB. Unfortunately we can't switch everything here to string_views as some
routines require functionality that's not avaialble for string_views. Namely
anything that converts a string to a numeric value. The story here is a bit sad
even in C++17 with std string_views.