Remove runtime selection of rng engines
This feature has in my opinion questionable user value. I'd be surprised if
anyone actually uses it on a regular basis.

This lets us devirtualize the engine with all the related benefits:
* no pointer inderiction on access (we can embed it right into the sim object)
* no virtual dispatch

The setup with `basic_rng_t` template is somewhat overengineered to handle
compile-time selection of rng engines and to actually let us test & compare the
engines in the "unit test". It can also be extended to bring back support for
runtime engine selection if the need arises.

Unsuprisingly it speeds up the sim a bit (with all of the usual caveats of
machine/compiler/etc.):

  nuohep@r2d2:~/dev/simc> for i in `seq 1 500`; do \
  > for exec in 'rng' 'bfa-dev'; do \
  > ./build-release/simc.$exec profiles/Tier25/T25_Hunter_Beast_Mastery.simc | \
  >   rg -o -r '$1' -e 'WallSeconds\s*=\s*([0-9.]+)' >> $exec-patchwerk.csv ; \
  > done; done
  nuohep@r2d2:~/dev/simc> ministat -q bfa-dev-patchwerk.csv rng-patchwerk.csv
  x bfa-dev-patchwerk.csv
  + rng-patchwerk.csv
      N           Min           Max        Median           Avg        Stddev
  x 500     1.3750636      2.145222     1.5516013     1.5675446   0.071085515
  + 500     1.3329705     2.2513635     1.5070543     1.5233121   0.074769679
  Difference at 95.0% confidence
	  -0.0442325 +/- 0.00904308
	  -2.82177% +/- 0.576895%
	  (Student's t, pooled s = 0.0729509)

While at it, further cleanup test code.