Initial structure for adding report level support.
Add a parse_reports option function which should parse all options beginning with eg. json, and with an optional postfix report level. So json2= or json7=

json= is equal to json=1, so the starting level of a report is 1 and not 0. (Following existing practice for the json report).

The parsed report entries are added to a vector, which is then used at the end of the simulation to generate the reports. This also allows to generate multiple reports of the same type.

Current implementation works only for json report.